Refactored SLURM account test to v2
