+++ tests/v2/run_mock_tasks.spec.js
@@ -1,14 +1,10 @@
 import { expect, test } from './workflow_fixture.js';
 import { waitModalClosed, waitPageLoading } from '../utils.js';
-import path from 'path';
-import os from 'os';
 import fs from 'fs';
 import { createDataset } from './dataset_utils.js';
+import { waitTaskSubmitted, waitTasksSuccess } from './workflow_task_utils.js';
 
-test('Collect and run mock tasks [v2]', async ({ page, workflow, request }) => {
-	const tasksMockWheelUrl =
-		'https://github.com/fractal-analytics-platform/fractal-server/raw/main/tests/v2/fractal_tasks_mock/dist/fractal_tasks_mock-0.0.1-py3-none-any.whl';
-
+test('Run mock tasks [v2]', async ({ page, workflow }) => {
 	await page.waitForURL(workflow.url);
 	await waitPageLoading(page);
 
@@ -30,91 +26,6 @@ test('Collect and run mock tasks [v2]', async ({ page, workflow, request }) => {
 		const { name } = await createDataset(page, workflow.projectId);
 		datasetName2 = name;
 	});
-
-	await test.step('Go to tasks page', async () => {
-		await page.goto('/v2/tasks');
-		await waitPageLoading(page);
-	});
-
-	let skipTasksCollection = false;
-	if ((await page.getByRole('table').last().getByText('MIP_compound').count()) > 0) {
-		console.warn('WARNING: Mock tasks V2 already collected. Skipping tasks collection');
-		skipTasksCollection = true;
-	}
-
-	if (!skipTasksCollection) {
-		/** @type {string} */
-		let tasksMockWheelFile;
-
-		await test.step('Download fractal_tasks_mock wheel', async () => {
-			const response = await request.get(tasksMockWheelUrl);
-			expect(response.status()).toEqual(200);
-			const body = await response.body();
-			const tmpDir = path.resolve(os.tmpdir(), 'playwright');
-			if (!fs.existsSync(tmpDir)) {
-				fs.mkdirSync(tmpDir);
-			}
-			tasksMockWheelFile = path.resolve(tmpDir, 'fractal_tasks_mock-0.0.1-py3-none-any.whl');
-			fs.writeFileSync(tasksMockWheelFile, body);
-		});
-
-		await test.step('Collect mock tasks', async () => {
-			await page.getByText('Local', { exact: true }).click();
-			await page.getByRole('textbox', { name: 'Package', exact: true }).fill(tasksMockWheelFile);
-			await page.getByRole('button', { name: 'Collect', exact: true }).click();
-
-			if (await page.getByText('Already installed').isVisible()) {
-				skipTasksCollection = true;
-				return;
-			}
-
-			// Wait for Task collections table
-			await page.waitForFunction(
-				(expectedCount) => document.querySelectorAll('table').length === expectedCount,
-				2
-			);
-			await expect(
-				page.locator('table tbody tr:first-child td:nth-child(2)').first()
-			).toContainText('fractal_tasks_mock');
-			expect(await getStatus(page)).toMatch(/^(pending|installing)$/);
-		});
-
-		if (!skipTasksCollection) {
-			await test.step('Wait tasks collection', async () => {
-				await page.waitForFunction(
-					() =>
-						document.querySelector('table tbody tr:first-child td:nth-child(4)')?.textContent ===
-						'OK'
-				);
-			});
-
-			await test.step('Check tasks list', async () => {
-				await page.getByRole('table').last().getByRole('cell', { name: 'MIP_compound' }).waitFor();
-			});
-
-			await test.step('Delete task collection log', async () => {
-				const deleteCollectionLogBtn = page.locator(
-					'table tr:first-child td:nth-child(5) button.btn-warning'
-				);
-				await deleteCollectionLogBtn.click();
-
-				// Confirm action modal
-				const modalTitle = page.locator('.modal.show .modal-title');
-				await modalTitle.waitFor();
-				await expect(modalTitle).toHaveText('Confirm action');
-				await expect(page.locator('.modal.show .modal-body')).toContainText(
-					'Remove a task collection log'
-				);
-
-				// Confirm the deletion
-				await page.getByRole('button', { name: 'Confirm' }).click();
-			});
-		}
-
-		await test.step('Cleanup temporary wheel file', async () => {
-			fs.rmSync(tasksMockWheelFile);
-		});
-	}
 
 	await test.step('Go to workflow page', async () => {
 		await page.goto(workflow.url);
@@ -402,15 +313,6 @@ test('Collect and run mock tasks [v2]', async ({ page, workflow, request }) => {
 
 /**
  * @param {import('@playwright/test').Page} page
- * @return {Promise<string>}
- */
-async function getStatus(page) {
-	const statusCell = page.locator('table tbody tr:first-child td:nth-child(4)').first();
-	return await statusCell.innerText();
-}
-
-/**
- * @param {import('@playwright/test').Page} page
  * @param {import('@playwright/test').Locator} row
  * @returns {Promise<string>}
  */
@@ -423,34 +325,4 @@ async function getJobId(page, row) {
 	expect(match).not.toBeNull();
 	await page.locator('.modal.show .modal-header [aria-label="Close"]').click();
 	return /** @type {RegExpMatchArray} */ (match)[1];
-}
-
-/**
- * @param {import('@playwright/test').Page} page
- * @param {number} count
- */
-async function waitTaskSubmitted(page, count) {
-	await expect(page.getByText('The last job failed with the following error')).toHaveCount(0);
-	if ((await page.locator('.job-status-icon.bi-check').count()) > 0) {
-		return;
-	}
-	const spinners = page.locator('.job-status-submitted.spinner-border');
-	await expect(spinners).toHaveCount(count);
-}
-
-/**
- * @param {import('@playwright/test').Page} page
- * @param {number} count
- */
-async function waitTasksSuccess(page, count) {
-	const spinners = page.locator('.job-status-submitted.spinner-border');
-	await expect(spinners).toHaveCount(0);
-	const errorAlert = page.getByText('The last job failed with the following error');
-	if (await errorAlert.isVisible()) {
-		const error = await page.locator('.alert.border-danger').innerText();
-		console.error(error);
-	}
-	await expect(errorAlert).toHaveCount(0);
-	await page.locator('.job-status-icon.bi-check').first().waitFor();
-	await expect(page.locator('.job-status-icon.bi-check')).toHaveCount(count);
 }
